Build 31012301

+++ ESP32/DIY-Flow-Bench/mafData/ACDELCO_92281162.cpp
@@ -72,7 +72,7 @@ int Maf::mafUnits() {
  * @note Global vector of vectors containing MAF>Flow key>value pairs
  ***/
 // std::vector<std::vector<int>> Maf::mafLookupTable = new std::vector<std::vector<int>> {{ // declare on heap using pointer
- std::vector<std::vector<int>> mafLookupTable = {{ 
+ std::vector<std::vector<long>> mafLookupTable = {{ 
     {0,0},
     {1,177},
     {2,181},
